pull will only remove file if it exists

+++ pull.sh
@@ -1,8 +1,12 @@
 git reset --hard && git pull
 
-rm ~/.bash_aliases
+if [ -f ~/.bash_aliases ]; then
+    rm ~/.bash_aliases
+fi
 cp ./configuration/.bash_aliases ~/.bash_aliases 
 
-rm ~/.bash_functions
+if [ -f ~/.bash_functions ]; then
+    rm ~/.bash_functions
+fi
 cp ./configuration/.bash_functions ~/.bash_functions
